Remove module `Location.Internal`.
There's no longer a need to have a separate internal module for
the `Location` type.
