+++ roc-id.cabal
@@ -74,7 +74,6 @@ library
       ROC.ID.Location
       ROC.ID.Nationality
   other-modules:
-      ROC.ID.Location.Internal
       ROC.ID.Unchecked
       ROC.ID.Utilities
   hs-source-dirs:
